Test with chsh command

+++ install.sh
@@ -1,13 +1,20 @@
 #!/bin/bash
 # Install oh-my-zsh
-sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+[ ! -d "$HOME/.oh-my-zsh" ] && sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
 # Install starship shell
-curl -sS https://starship.rs/install.sh | sh
+if ! [ -x "$(command -v starship)" ]; then
+  curl -sS https://starship.rs/install.sh | sh
+fi
+
 
 # Install fzf
-git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
-~/.fzf/install
+if ! [ -x "$(command -v fzf)" ]; then
+  git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+  ~/.fzf/install
+fi
 
 # Symlink configs
-ln -s ./zshrc ~/.zshrc
+[ ! -L "$HOME/.zshrc" ] && ln -s ./zshrc ~/.zshrc
+
+chsh -s "$(which zsh)"
